Changes from Wenqin (wqxu@ucla.edu):
For the cross talk part, make the energy assignments of the newly created
strips happen before they are added to the mEmcMcHits collection, otherwise
the energy assignments won't be recorded in the collection.

The cross talk leakage is now proportional to the difference of energies of two neighbor
strips. The leakage range is extended to +/- 5 strips.

The cross talk in the BSMD-Phi still needs more solid adjustment, disabled for now.